move encryption to models/utils
